+++ handpose/helpers.py
@@ -1,5 +1,7 @@
 import yaml
 import torch
+import os
+import re
 
 def load_variables(file_path):
     """Loads variables from a YAML file.
@@ -368,3 +370,46 @@ def detect(image, model, require_kpt_conf, S, B, nkpt, nc, iou_threshold):
     head_nms = non_max_suppression(head)
 
     return head, head_nms
+
+def create_train_dir(save_model_path):
+    """Creates a directory for training results.
+
+    Parameters
+    ----------
+    save_model_path: str
+        Path to store weights.
+        
+    """
+    if not os.path.exists(save_model_path):
+        os.makedirs(save_model_path)
+    
+    dirs = os.listdir(save_model_path)
+
+    # If the directory is empty hard code to create 'train1'
+    if not dirs:
+        train_dir_name = os.path.join(save_model_path, 'train1')
+        os.makedirs(train_dir_name)
+    else:
+        train_nums = []
+
+        for d in dirs:
+            d_train = re.findall('train\d+', d)
+            if d_train:
+                train_nums.append(int(re.findall('\d+', d)[0]))
+
+        # If train_nums is empty list i.e. no train# was found
+        # Hard code to create 'train1'
+        if not train_nums:
+            train_dir_name = os.path.join(save_model_path, 'train1')
+            os.makedirs(train_dir_name)
+        else:
+            train_nums.sort()
+
+            # Removes the last number from train_nums
+            last_train_num = train_nums.pop()
+
+            train_dir_name = os.path.join(save_model_path, 'train' + str(last_train_num + 1))
+
+            os.makedirs(train_dir_name)
+
+    return train_dir_name