+++ handpose/train.py
@@ -7,6 +7,7 @@ import torch.optim as optim
 from torch.optim import lr_scheduler
 import pickle
 import pandas as pd
+from pathlib import Path
 
 sys.path.append('../')
 from handpose import *
@@ -37,7 +38,8 @@ def train_model(dataloaders,
                 scheduler='default',
                 individual_plots=False,
                 losses_types=["total_loss", "box_loss", "conf_loss", "class_loss", "kpt_loss"],
-                sample_grid_shape=(2,4)
+                sample_grid_shape=(2,4),
+                **kwargs
                ):
     r"""Training function.
     
@@ -108,15 +110,27 @@ def train_model(dataloaders,
     
     """
 
+    if kwargs:
+        resume_training = kwargs['resume_training']
+        loss_df = kwargs['loss_df']
+        train_path = kwargs['train_save_path']
+        epochs_passed, _ = loss_df['train'].shape
+
+    else:
+        resume_training = False
+        epochs_passed = 0
+        loss_df = {'train': pd.DataFrame(), 'valid': pd.DataFrame()}
+
     # Record start time
     since = time.time()
 
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
     if save_model_path:
-        timestamp = str(int(since))
-        train_dir = train_dir + '_' + timestamp
-        train_path = os.path.join(save_model_path, train_dir)
+        if not resume_training:
+            # Creating training directory
+            train_path = create_train_dir(save_model_path)
+            print(f"Created a new directory at {train_path}")
 
         # Last model path to save model after every epoch
         last_model_path = os.path.join(train_path, 'last.pt')
@@ -131,6 +145,12 @@ def train_model(dataloaders,
     valid_losses = copy.deepcopy(losses)
 
     all_losses = {'train': losses, 'valid': valid_losses}
+
+    # If resume training is true populate the previous losses in the all_losses dict
+    if resume_training:
+        for phase in all_losses:
+            for c in loss_df[phase].columns:
+                all_losses[phase][c] = list(loss_df[phase][c])
     
     eval_metrics = dict()
 
@@ -148,7 +168,7 @@ def train_model(dataloaders,
         scheduler = Scheduler(optimizer, lr_scheduler.StepLR ,**{'step_size': 7, 'gamma': 0.1}).scheduler
         
     
-    for epoch in tqdm(range(num_epochs), unit='batch', total=num_epochs):
+    for epoch in tqdm(range(epochs_passed, num_epochs), unit='batch', total=num_epochs):
         if verbose:
             print('\n\n')
             print('=' * len(f"Epoch: {epoch + 1}"))
@@ -257,26 +277,27 @@ def train_model(dataloaders,
                 epoch_losses[k] = epoch_loss
                 all_losses[phase][k].append(epoch_loss)
 
+            # Append all_losses to dataframe here
+            loss_df[phase] = loss_df[phase].from_dict(all_losses[phase])
+
+            if save_model_path:
+                loss_df[phase].to_csv(os.path.join(train_path, f'{phase}_loss.csv'), index=False)
+
             mAP_epoch = mAP / dataset_sizes[phase]
             epochs_mAP[phase].append(mAP_epoch)
 
             if verbose:
+                print('\n')
                 print(f'{phase}')
                 print('-' * len(str(phase)))
-                for k,v in epoch_losses.items():
-                    print(f'{k}: {v:.3f}')
-                print("\n")
 
-                # printing mAP
-                print("Metrics")
-                print('-' * len("Metrics"))
-                print(f"mAP: {mAP_epoch}")
+                df_metrics = loss_df[phase].iloc[-1:].copy()
+
+                df_metrics.loc[:, f'mAP{int(iou_threshold * 100)}'] = [mAP_epoch]
+
+                print(df_metrics)
 
             if save_model_path:
-                if not os.path.exists(train_path):
-                    os.makedirs(train_path)
-                    print(f'New directory {train_dir} created at {train_path}')
-
                 torch.save(model.state_dict(), last_model_path)
 
     time_elapsed = time.time() - since
@@ -311,14 +332,9 @@ def train_model(dataloaders,
 
             plot_single_history(mAP_plot_history, root_path=train_path, save_path_prefix='', xlabel='Epoch', ylabel='mAP')
 
-        plot_all_history(loss_history, root_path=train_path)
+        # plot_all_history(loss_history, root_path=train_path)
 
-        # CSV files
-        train_loss_df = pd.DataFrame(plot_history['train'])
-        valid_loss_df = pd.DataFrame(plot_history['valid'])
-
-        train_loss_df.to_csv(os.path.join(train_path, 'train_loss.csv'), index=False)
-        valid_loss_df.to_csv(os.path.join(train_path, 'valid_loss.csv'), index=False)
+        plot_loss(loss_df, root_path=train_path)
 
         # Save sample images
         save_sample_images(images, 
@@ -440,4 +456,143 @@ def train_network(config, verbose=True):
                 learning_rate=learning_rate,
                 lr_momentum=lr_momentum,
                 scheduler=scheduler
+               )
+
+def resume_training(train_dir, config_file, verbose=True):
+    """Restarts training from the last checkpoint"""
+
+    # Sanity check - Make sure the required files are present
+    
+    # last.pt
+    try:
+        model_path = os.path.join(train_dir, 'last.pt')
+        if not Path(model_path).is_file():
+            print(f"{model_path} not found!")
+            sys.exit(1)
+    except Exception as e:
+        print(f'{e}')
+    
+    # `train_loss.csv` and `valid_loss.csv`
+    loss_df = dict()
+    
+    try:
+        loss_df['train'] = pd.read_csv(os.path.join(train_dir, 'train_loss.csv'))
+        loss_df['valid'] = pd.read_csv(os.path.join(train_dir, 'valid_loss.csv'))
+    except Exception as e:
+        print(f'{e}')
+
+    # Config file 
+    config = load_variables(config_file)
+
+    try:
+        REPO = config['REPO']
+        S = config['S']
+        B = config['B']
+        nkpt = config['nkpt']
+        nkpt_dim = config['nkpt_dim']
+        nc = config['nc']
+        batch_size = config['batch_size']
+        cell_relative = config['cell_relative']
+        input_size = tuple(config['input_size'])
+        require_kpt_conf = config['require_kpt_conf']
+        weights = config['weights']
+        model_name = config['model_name']
+        freeze_weights = config['freeze_weights']
+        data_dir = config['data_dir']
+        save_model_path = config['save_model_path']
+        num_epochs = config['num_epochs']
+        iou_threshold = config['iou_threshold']
+        lambda_coord = config['lambda_coord']
+        lambda_noobj = config['lambda_noobj']
+        epsilon = float(config['epsilon'])
+        lambda_kpt = config['lambda_kpt']
+        lambda_kpt_conf = config['lambda_kpt_conf']
+        shuffle_data = config['shuffle_data']
+        num_workers = config['num_workers']
+        drop_last = config['drop_last']
+        optimizer = config['optimizer']
+        learning_rate = config['learning_rate']
+        lr_momentum = config['lr_momentum']
+        scheduler = config['scheduler']
+    except Exception as e:
+        print(f"{e}")
+
+    # Dataloaders
+    dataloaders, dataset_sizes = get_dataloaders(data_dir,
+                                             S,
+                                             nc,
+                                             nkpt,
+                                             cell_relative,
+                                             require_kpt_conf,
+                                             batch_size,
+                                             shuffle_data,
+                                             num_workers,
+                                             drop_last
+                                            )
+
+    # Device
+    DEVICE = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+
+    model = TransferNetwork(
+        repo_or_dir=REPO,
+        model_name=model_name,
+        weights=weights,
+        S=S,
+        B=B,
+        nkpt=nkpt,
+        nc=nc,
+        input_size=input_size,
+        require_kpt_conf=require_kpt_conf,
+        freeze_weights=freeze_weights
+    ).to(DEVICE)
+
+    # Reload the saved model
+    model.load_state_dict(torch.load(model_path, map_location=DEVICE))
+
+    # Counts the epochs passed from the epochs given by the rows of 
+    epochs_passed, _ = loss_df['train'].shape
+
+    # sanity check on num_epochs
+    if num_epochs <=0:
+        print("""Cannot perform training on epochs <= 0! 
+        Make sure the config file num_epochs is larger value or start a new training by preloading weights.
+        """)
+        sys.exit(1)
+
+    print(model.summary())
+
+    print(f"""
+    \n
+    Restarting training from {epochs_passed + 1}.
+    Using weights from {model_path}.
+    """)
+
+    retraining_args = {'train_save_path': train_dir,
+                       'resume_training': True,
+                       'loss_df': loss_df
+                      }
+
+    history = train_model(dataloaders,
+                dataset_sizes,
+                model,
+                loss_fn,   
+                num_epochs=num_epochs,
+                S=S,
+                B=B,
+                nkpt=nkpt,
+                nc=nc,
+                require_kpt_conf=require_kpt_conf,
+                iou_threshold=iou_threshold,
+                lambda_coord=lambda_coord,
+                lambda_noobj=lambda_noobj,
+                epsilon=epsilon,
+                lambda_kpt=lambda_kpt,
+                lambda_kpt_conf=lambda_kpt_conf,
+                verbose=verbose,
+                save_model_path=save_model_path,
+                optimizer=optimizer,
+                learning_rate=learning_rate,
+                lr_momentum=lr_momentum,
+                scheduler=scheduler,
+                **retraining_args
                )